pyproject.toml
[project]
name = "cosmos-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.14.0",
    "azure-identity>=1.25.1",
    "azure-mgmt-authorization>=4.0.0",
    "azure-mgmt-cosmosdb>=9.8.0",
    "msgraph-core>=1.3.8",
    "msgraph-sdk>=1.47.0",
    "python-dotenv>=1.2.1",
]
